# openapi.yaml
openapi: 3.1.0
info:
  title: Maru API
  version: 0.1.0
  description: Versioned API for the Maru convention operating platform.
paths:
...
  /api/v1/me/restrictions:
    get:
      operationId: identity_list_my_restrictions
      tags:
      - me
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountRestriction'
          description: ''
  /api/v1/me/restrictions/{restriction_id}/appeals:
    post:
      operationId: identity_appeal_my_restriction
      parameters:
      - in: path
        name: restriction_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - me
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestrictionAppealCreate'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestrictionAppeal'
          description: ''
  /api/v1/me/security-history:
    get:
      operationId: identity_list_my_security_history
      tags:
      - me
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountSecurityEvent'
          description: ''
  /api/v1/me/sessions:
    get:
      operationId: identity_list_my_sessions
      tags:
      - me
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountSession'
          description: ''
  /api/v1/me/sessions/{session_id}/revoke:
    post:
      operationId: identity_revoke_my_session
      parameters:
      - in: path
        name: session_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - me
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/me/step-up:
    post:
      operationId: identity_complete_step_up
      tags:
      - me
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepUp'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/v1/organizations/{organization_id}/editions/{edition_id}/restriction-appeals/{appeal_id}/decision:
    post:
      operationId: identity_decide_restriction_appeal
      parameters:
      - in: path
        name: appeal_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: edition_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - organizations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffRestrictionAppealDecision'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestrictionAppeal'
          description: ''
  /api/v1/organizations/{organization_id}/editions/{edition_id}/restrictions:
    get:
      operationId: identity_list_scoped_restrictions
      parameters:
      - in: path
        name: edition_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - organizations
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountRestriction'
          description: ''
    post:
      operationId: identity_issue_scoped_restriction
      parameters:
      - in: path
        name: edition_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - organizations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffRestrictionCreate'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountRestriction'
          description: ''
  /api/v1/organizations/{organization_id}/editions/{edition_id}/restrictions/{restriction_id}/revoke:
    post:
      operationId: identity_revoke_scoped_restriction
      parameters:
      - in: path
        name: edition_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: restriction_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - organizations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffRestrictionRevoke'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountRestriction'
          description: ''
...
  /api/v1/public/accounts:
    post:
      operationId: identity_bootstrap_account
      tags:
      - public
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountBootstrap'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/public/accounts/recovery:
    post:
      operationId: identity_request_recovery
      tags:
      - public
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoveryRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/public/accounts/recovery/complete:
    post:
      operationId: identity_complete_recovery
      tags:
      - public
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoveryComplete'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/public/accounts/verify-email:
    post:
      operationId: identity_verify_email
      tags:
      - public
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Token'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/public/csrf:
    get:
      operationId: identity_get_csrf_token
      tags:
      - public
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/v1/public/sessions:
    post:
      operationId: identity_create_public_session
      tags:
      - public
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionSignIn'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
components:
  schemas:
...
    AccountRestriction:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        organization_id:
          type: string
          format: uuid
          readOnly: true
        edition_id:
          type:
          - string
          - 'null'
          format: uuid
        kind:
          allOf:
          - $ref: '#/components/schemas/AccountRestrictionKindEnum'
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/ActiveRevokedStatusEnum'
          readOnly: true
        attendee_message:
          type: string
          readOnly: true
        effective_at:
          type: string
          format: date-time
          readOnly: true
        expires_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
        revoked_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
        appeal_status:
          type:
          - string
          - 'null'
          readOnly: true
      required:
      - appeal_status
      - attendee_message
      - edition_id
      - effective_at
      - expires_at
      - id
      - kind
      - organization_id
      - revoked_at
      - status
...
    AccountSecurityEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        event_type:
          allOf:
          - $ref: '#/components/schemas/EventTypeEnum'
          readOnly: true
        event_label:
          type: string
        outcome:
          allOf:
          - $ref: '#/components/schemas/AccountSecurityEventOutcomeEnum'
          readOnly: true
        occurred_at:
          type: string
          format: date-time
          readOnly: true
        source_channel:
          type: string
          readOnly: true
        detail_code:
          type: string
          readOnly: true
      required:
      - detail_code
      - event_label
      - event_type
      - id
      - occurred_at
      - outcome
      - source_channel
...
    AccountSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        label:
          type: string
          readOnly: true
        created_channel:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        last_seen_at:
          type: string
          format: date-time
          readOnly: true
        step_up_verified_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
        active:
          type: boolean
        revoked_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
        revocation_reason:
          type: string
          readOnly: true
      required:
      - active
      - created_at
      - created_channel
      - id
      - label
      - last_seen_at
      - revocation_reason
      - revoked_at
      - step_up_verified_at
...
    RecoveryRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
      - email
...
    RestrictionAppeal:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        restriction_id:
          type: string
          format: uuid
          readOnly: true
        statement:
          type: string
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/RestrictionAppealStatusEnum'
          readOnly: true
        submitted_at:
          type: string
          format: date-time
          readOnly: true
        decided_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
        decision_summary:
          type: string
          readOnly: true
      required:
      - decided_at
      - decision_summary
      - id
      - restriction_id
      - statement
      - status
      - submitted_at
    RestrictionAppealCreate:
      type: object
      properties:
        statement:
          type: string
          maxLength: 4000
      required:
      - statement
...
    StaffRestrictionAppealDecision:
      type: object
      properties:
        decision:
          $ref: '#/components/schemas/StaffRestrictionAppealDecisionDecisionEnum'
        summary:
          type: string
          maxLength: 500
      required:
      - decision
      - summary
...
    StaffRestrictionCreate:
      type: object
      properties:
        account_id:
          type: string
          format: uuid
        kind:
          $ref: '#/components/schemas/AccountRestrictionKindEnum'
        reason_code:
          type: string
          maxLength: 80
          pattern: ^[-a-zA-Z0-9_]+$
        attendee_message:
          type: string
          maxLength: 320
        internal_reference:
          type: string
          maxLength: 120
        effective_at:
          type: string
          format: date-time
        expires_at:
          type:
          - string
          - 'null'
          format: date-time
        notify_account:
          type: boolean
          default: true
      required:
      - account_id
      - attendee_message
      - effective_at
      - kind
      - reason_code
    StaffRestrictionRevoke:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
      required:
      - reason
...
    StepUp:
      type: object
      properties:
        password:
          type: string
          writeOnly: true
          maxLength: 128
      required:
      - password
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid